refactor#131 rename bottomsheet
